server/requirements.macos.m1.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
darker==1.5.1
dj-database-url==1.0.0
django-extensions==3.2.0
django-model-utils==4.2.0
Django==3.2.15
ebird-api==3.0.6
gTTS==2.2.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pydub==0.25.1
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
